backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
django-jsonview==2.0.0
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3